Utilize DiskStorage
This CLI refactor is effectively a no-op. The purpose here was to
replace any filesystem I/O that directly occurs in any CLI command file.
Eventually, this will be refactored again to enable the sharding of
instance-specific files such as output and ledger data.

1. yarn && yarn build
2. cd sharness/__snapshots__/test-instance
3. scdev go
4. scdev serve

Verify that all data loads successfully in the webpage.